requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.1.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
environs==8.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
marshmallow==3.26.1
multidict==6.1.0
packaging==24.2
propcache==0.3.0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
